Fix an issue with initial scale type

+++ lib/src/main/java/com/soundcloud/android/crop/CropImageActivity.java
@@ -261,7 +261,7 @@ public class CropImageActivity extends MonitoredActivity {
             private int height;
             private int cropWidth;
             private int cropHeight;
-            private static final float SCALE_FACTOR = 0.9;
+            private static final double SCALE_FACTOR = 0.9;
 
             public CropZoneOptimizer(int width, int height, int cropWidth, int cropHeight) {
                 this.width = width;
